fix import issues caused by refactor
